+++ neoforge/src/main/java/yesman/epicfight/api/event/types/player/ModifyComboCounter.java
@@ -1,10 +1,19 @@
 package yesman.epicfight.api.event.types.player;
 
+import com.google.common.collect.Sets;
+import org.jetbrains.annotations.Nullable;
 import yesman.epicfight.api.animation.AnimationManager.AnimationAccessor;
+import yesman.epicfight.api.animation.types.AttackAnimation;
+import yesman.epicfight.api.animation.types.MainFrameAnimation;
 import yesman.epicfight.api.animation.types.StaticAnimation;
 import yesman.epicfight.api.event.CancelableEvent;
 import yesman.epicfight.api.event.LivingEntityPatchEvent;
+import yesman.epicfight.world.capabilities.entitypatch.player.PlayerPatch;
 import yesman.epicfight.world.capabilities.entitypatch.player.ServerPlayerPatch;
+import yesman.epicfight.world.capabilities.item.CapabilityItem;
+
+import java.util.List;
+import java.util.Set;
 
 public class ModifyComboCounter extends LivingEntityPatchEvent implements CancelableEvent {
 	private final Causal causal;
@@ -48,4 +57,40 @@ public class ModifyComboCounter extends LivingEntityPatchEvent implements Cancel
 	public enum Causal {
 		ANOTHER_ACTION_ANIMATION, TIME_EXPIRED
 	}
+
+    @FunctionalInterface
+    public interface ComboCounterHandler {
+        ComboCounterHandler DEFAULT_COMBO_HANDLER = (CapabilityItem itemCapability, Causal causal, PlayerPatch<?> entitypatch, @Nullable AnimationAccessor<? extends MainFrameAnimation> nextAnimation, int comboCounter) -> {
+            // When causal is time expiration, reset the counter
+            if (causal == Causal.TIME_EXPIRED) {
+                return 0;
+            }
+
+            List<AnimationAccessor<? extends AttackAnimation>> comboAnimations = itemCapability.getAutoAttackMotion(entitypatch);
+
+            if (comboAnimations == null) {
+                return 0;
+            }
+
+            Set<AnimationAccessor<? extends AttackAnimation>> attackMotionSet = Sets.newHashSet(comboAnimations);
+
+            // when the next animation is not included in weapon attack motion, reset the counter
+            if (!attackMotionSet.contains(nextAnimation)) {
+                return 0;
+            }
+
+            int comboSize = comboAnimations.size();
+
+            // When the next animation is dash or air attacks, reset combo counter
+            if (nextAnimation.equals(comboAnimations.get(comboSize - 1)) || nextAnimation.equals(comboAnimations.get(comboSize - 2))) {
+                return 0;
+            }
+
+            return comboCounter;
+        };
+
+        /// Control the combo counter with given parameters
+        /// @param nextAnimation null when ComboCounterHandleEvent.Causal is [Causal#TIME_EXPIRED]
+        int handleComboCounter(CapabilityItem itemCapability, Causal causal, PlayerPatch<?> entitypatch, @Nullable AnimationAccessor<? extends MainFrameAnimation> nextAnimation, int comboCounter);
+    }
 }